backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
openai
httpx
python-multipart
reportlab
twilio
python-jose[cryptography]
passlib
bcrypt==4.0.1
